refactored to use a shared subflow for create and update
