+++ contracts/src/main/kotlin/com/r3/developers/csdetemplate/utxoexample/states/ChatState.kt
@@ -4,7 +4,6 @@ import com.r3.developers.csdetemplate.utxoexample.contracts.ChatContract
 import net.corda.v5.base.types.MemberX500Name
 import net.corda.v5.ledger.utxo.BelongsToContract
 import net.corda.v5.ledger.utxo.ContractState
-import net.corda.v5.membership.MemberInfo
 import java.security.PublicKey
 import java.util.*
