+++ workflows/src/main/kotlin/com/r3/developers/csdetemplate/utxoexample/workflows/AppendChatSubFlow.kt
@@ -1,0 +1,71 @@
+package com.r3.developers.csdetemplate.utxoexample.workflows
+
+import com.r3.developers.csdetemplate.utxoexample.states.ChatState
+import net.corda.v5.application.flows.*
+import net.corda.v5.application.messaging.FlowMessaging
+import net.corda.v5.application.messaging.FlowSession
+import net.corda.v5.base.annotations.Suspendable
+import net.corda.v5.base.types.MemberX500Name
+import net.corda.v5.base.util.contextLogger
+import net.corda.v5.ledger.utxo.UtxoLedgerService
+import net.corda.v5.ledger.utxo.transaction.UtxoSignedTransaction
+
+@InitiatingFlow(protocol = "append-chat-protocol")
+class AppendChatSubFlow(private val signedTransaction: UtxoSignedTransaction, private val otherMember: MemberX500Name): SubFlow<String> {
+
+    private companion object {
+        val log = contextLogger()
+    }
+
+    @CordaInject
+    lateinit var ledgerService: UtxoLedgerService
+
+    @CordaInject
+    lateinit var flowMessaging: FlowMessaging
+
+    @Suspendable
+    override fun call(): String {
+
+        log.info("AppendChatFlow.call() called")
+
+            val session = flowMessaging.initiateFlow(otherMember)
+
+            return try {
+                val finalizedSignedTransaction = ledgerService.finalize(
+                    signedTransaction,
+                    listOf(session)
+                )
+                finalizedSignedTransaction.id.toString().also {
+                    log.info("Success! Response: $it")
+                }
+            } catch (e: Exception) {
+                log.warn("Finality failed", e)
+                "Finality failed, ${e.message}"
+            }
+    }
+}
+
+@InitiatedBy("append-chat-protocol")
+class AppendChatResponderFlow: ResponderFlow {
+
+    private companion object {
+        val log = contextLogger()
+    }
+
+    @CordaInject
+    lateinit var utxoLedgerService: UtxoLedgerService
+
+    @Suspendable
+    override fun call(session: FlowSession) {
+        try {
+            val finalizedSignedTransaction = utxoLedgerService.receiveFinality(session) { ledgerTransaction ->
+                val state = ledgerTransaction.outputContractStates.first() as ChatState
+                if (checkForBannedWords(state.message) && checkMessageFromMatchesCounterparty(state, session.counterparty)) throw IllegalStateException("Failed verification")
+                log.info("Verified the transaction- ${ledgerTransaction.id}")
+            }
+            log.info("Finished responder flow - ${finalizedSignedTransaction.id}")
+        } catch (e: Exception) {
+            log.warn("Exceptionally finished responder flow", e)
+        }
+    }
+}
