+++ workflows/src/main/kotlin/com/r3/developers/csdetemplate/utxoexample/workflows/UpdateChatFlow.kt
@@ -6,9 +6,7 @@ import net.corda.v5.application.flows.*
 import net.corda.v5.application.marshalling.JsonMarshallingService
 import net.corda.v5.application.membership.MemberLookup
 import net.corda.v5.application.messaging.FlowMessaging
-import net.corda.v5.application.messaging.FlowSession
 import net.corda.v5.base.annotations.Suspendable
-import net.corda.v5.base.types.MemberX500Name
 import net.corda.v5.base.util.contextLogger
 import net.corda.v5.base.util.days
 import net.corda.v5.ledger.utxo.UtxoLedgerService
@@ -25,17 +23,11 @@ class UpdateChatFlow: RPCStartableFlow {
         val log = contextLogger()
     }
 
-    // JsonMarshallingService provides a Service for manipulating json
     @CordaInject
     lateinit var jsonMarshallingService: JsonMarshallingService
 
-    // MemberLookup provides a service for looking up information about members of the Virtual Network which
-    // this CorDapp is operating in.
     @CordaInject
     lateinit var memberLookup: MemberLookup
-
-    @CordaInject
-    lateinit var utxoLedgerService: UtxoLedgerService
 
     @CordaInject
     lateinit var ledgerService: UtxoLedgerService
@@ -43,21 +35,21 @@ class UpdateChatFlow: RPCStartableFlow {
     @CordaInject
     lateinit var flowMessaging: FlowMessaging
 
+    @CordaInject
+    lateinit var flowEngine: FlowEngine
+
     @Suspendable
     override fun call(requestBody: RPCRequestData): String {
 
         log.info("UpdateNewChatFlow.call() called")
 
         try {
-
             val flowArgs = requestBody.getRequestBodyAs(jsonMarshallingService, UpdateChatFlowArgs::class.java)
 
             // look up state (this is very inefficient)
-
             val stateAndRef = ledgerService.findUnconsumedStatesByType(ChatState::class.java).singleOrNull {
                 it.state.contractState.id == flowArgs.id
             } ?: throw Exception("Multiple or zero Chat states with id ${flowArgs.id} found")
-
 
             val myInfo = memberLookup.myInfo()
             val state = stateAndRef.state.contractState
@@ -69,7 +61,7 @@ class UpdateChatFlow: RPCStartableFlow {
 
             val newChatState = state.updateMessage(myInfo.name, flowArgs.message)
 
-            val txBuilder= utxoLedgerService.getTransactionBuilder()
+            val txBuilder= ledgerService.getTransactionBuilder()
                 .setNotary(stateAndRef.state.notary)
                 .setTimeWindowBetween(Instant.now(), Instant.now().plusMillis(1.days.toMillis()))
                 .addOutputState(newChatState)
@@ -80,55 +72,13 @@ class UpdateChatFlow: RPCStartableFlow {
             @Suppress("DEPRECATION")
             val signedTransaction = txBuilder.toSignedTransaction(myInfo.ledgerKeys.first())
 
-            val session = flowMessaging.initiateFlow(otherMember.name)
-
-
-            return try {
-                val finalizedSignedTransaction = utxoLedgerService.finalize(
-                    signedTransaction,
-                    listOf(session)
-                )
-                finalizedSignedTransaction.id.toString().also {
-                    log.info("Success! Response: $it")
-                }
-            } catch (e: Exception) {
-                log.warn("Finality failed", e)
-                "Finality failed, ${e.message}"
-            }
+            return flowEngine.subFlow(AppendChatSubFlow(signedTransaction, otherMember.name))
 
         } catch (e: Exception) {
             log.warn("Failed to process utxo flow for request body '$requestBody' because:'${e.message}'")
             throw e
         }
     }
-}
-
-
-// todo: Can create and update share a responder? or combine create/ update into one <- do this
-@InitiatedBy("update-chat-protocol")
-class UpdateChatResponderFlow: ResponderFlow {
-
-    private companion object {
-        val log = contextLogger()
-    }
-
-    @CordaInject
-    lateinit var utxoLedgerService: UtxoLedgerService
-
-    @Suspendable
-    override fun call(session: FlowSession) {
-        try {
-            val finalizedSignedTransaction = utxoLedgerService.receiveFinality(session) { ledgerTransaction ->
-                val state = ledgerTransaction.outputContractStates.first() as ChatState
-                if (checkForBannedWords(state.message) && checkMessageFromMatchesCounterparty(state, session.counterparty)) throw IllegalStateException("Failed verification")
-                log.info("Verified the transaction- ${ledgerTransaction.id}")
-            }
-            log.info("Finished responder flow - ${finalizedSignedTransaction.id}")
-        } catch (e: Exception) {
-            log.warn("Exceptionally finished responder flow", e)
-        }
-    }
-
 }
 
 /*
